Creation of a function scrape_cat_list which returns a list of each category page
